@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -31,6 +31,7 @@ dependencies:
   sqflite: ^2.0.0+3
   path: ^1.8.0
   encrypt: ^5.0.0
+  clipboard: ^0.1.3
 
 dev_dependencies:
   flutter_test:
