@@ -6,7 +6,7 @@ class ShowNote extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final NoteModel note =
-        ModalRoute.of(context).settings.arguments as NoteModel;
+        ModalRoute.of(context)!.settings.arguments as NoteModel;
 
     return Scaffold(
       appBar: AppBar(
@@ -15,12 +15,18 @@ class ShowNote extends StatelessWidget {
           IconButton(
             icon: Icon(Icons.delete_forever),
             onPressed: () {
-              DatabaseProvider.db.deleteNote(note.id);
-              Navigator.pushNamedAndRemoveUntil(
-                context,
-                "home",
-                (route) => false,
+              showDialog(
+                context: context,
+                builder: (_) {
+                  return EliminacionAlert(note: note);
+                },
               );
+            },
+          ),
+          IconButton(
+            icon: Icon(Icons.edit),
+            onPressed: () {
+              Navigator.pushNamed(context, 'editNote');
             },
           )
         ],
@@ -46,6 +52,41 @@ class ShowNote extends StatelessWidget {
           ],
         ),
       ),
+    );
+  }
+}
+
+class EliminacionAlert extends StatelessWidget {
+  final NoteModel note;
+
+  const EliminacionAlert({
+    required this.note,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text('HOla'),
+      content: Text('¿Estás seguro de eliminar esta nota?'),
+      actions: [
+        TextButton(
+          onPressed: () {
+            Navigator.of(context).pop();
+          },
+          child: Text('Cancelar'),
+        ),
+        TextButton(
+          onPressed: () {
+            DatabaseProvider.db.deleteNote(note.id!);
+            Navigator.pushNamedAndRemoveUntil(
+              context,
+              "notes",
+              (route) => false,
+            );
+          },
+          child: Text('Aceptar'),
+        ),
+      ],
     );
   }
 }
