@@ -1,3 +1,4 @@
+import 'package:clipboard/clipboard.dart';
 import 'package:flutter/material.dart';
 import 'package:notes_app/db/database_provider.dart';
 import 'package:notes_app/models/password_model.dart';
@@ -7,55 +8,106 @@ class ShowPassword extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final PasswordModel pass =
-        ModalRoute.of(context).settings.arguments as PasswordModel;
+        ModalRoute.of(context)!.settings.arguments as PasswordModel;
 
     final passProvider = PasswordProvider();
 
     return Scaffold(
       appBar: AppBar(
-        title: Text('Nota'),
+        title: Text('Contraseña'),
         actions: [
           IconButton(
             icon: Icon(Icons.delete_forever),
             onPressed: () {
-              DatabaseProvider.db.deletePassword(pass.id);
-              Navigator.pushNamedAndRemoveUntil(
-                context,
-                "passwords",
-                (route) => false,
+              showDialog(
+                context: context,
+                builder: (context) {
+                  return AlertaEliminacion(id: pass.id);
+                },
+                barrierDismissible: false,
               );
             },
-          )
+          ),
+          IconButton(
+            icon: Icon(Icons.edit),
+            onPressed: () {
+              Navigator.pushNamed(context, "editPassword");
+            },
+          ),
         ],
       ),
       body: Padding(
         padding: EdgeInsets.all(8.0),
         child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            Text(
-              pass.title,
-              style: TextStyle(
-                fontSize: 28.0,
-                fontWeight: FontWeight.bold,
+            Container(
+              width: double.infinity,
+              child: Text(
+                pass.title,
+                textAlign: TextAlign.center,
+                style: TextStyle(
+                  fontSize: 28.0,
+                  fontWeight: FontWeight.bold,
+                ),
               ),
             ),
             SizedBox(height: 16.0),
-            Text(
-              pass.user,
-              style: TextStyle(
-                fontSize: 18.0,
-              ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  pass.user ?? '',
+                  style: TextStyle(
+                    fontSize: 18.0,
+                  ),
+                ),
+                IconButton(
+                  icon: Icon(Icons.copy),
+                  onPressed: () {
+                    FlutterClipboard.copy(pass.user ?? '').then(
+                      (value) {
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          SnackBar(
+                            content: Text('Usuario copiado'),
+                            duration: Duration(seconds: 1),
+                          ),
+                        );
+                      },
+                    );
+                  },
+                ),
+              ],
+            ),
+            SizedBox(height: 16.0),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  passProvider.decrypt(pass.password),
+                  style: TextStyle(
+                    fontSize: 18.0,
+                  ),
+                ),
+                IconButton(
+                  icon: Icon(Icons.copy),
+                  onPressed: () {
+                    FlutterClipboard.copy(passProvider.decrypt(pass.password))
+                        .then(
+                      (value) => ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text('Contraseña copiada'),
+                          duration: Duration(seconds: 1),
+                        ),
+                      ),
+                    );
+                  },
+                ),
+              ],
             ),
             SizedBox(height: 16.0),
             Text(
-              passProvider.decrypt(pass.password),
-              style: TextStyle(
-                fontSize: 18.0,
-              ),
-            ),
-            SizedBox(height: 16.0),
-            Text(
-              pass.description,
+              pass.description ?? '',
               style: TextStyle(
                 fontSize: 18.0,
               ),
@@ -63,6 +115,41 @@ class ShowPassword extends StatelessWidget {
           ],
         ),
       ),
+    );
+  }
+}
+
+class AlertaEliminacion extends StatelessWidget {
+  final id;
+
+  const AlertaEliminacion({
+    required this.id,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text('HOla'),
+      content: Text('¿Estás seguro de eliminar esta contraseña?'),
+      actions: [
+        TextButton(
+          onPressed: () {
+            Navigator.of(context).pop();
+          },
+          child: Text('Cancelar'),
+        ),
+        TextButton(
+          onPressed: () {
+            DatabaseProvider.db.deletePassword(id);
+            Navigator.pushNamedAndRemoveUntil(
+              context,
+              "passwords",
+              (route) => false,
+            );
+          },
+          child: Text('Aceptar'),
+        ),
+      ],
     );
   }
 }
